Add support for interfacing with Clang Blocks
Clang Blocks, or just Blocks [1], is an extension to the C family of
languages which adds support for anonymous functions, much like
delegates in D.

Being able to interface with Blocks is important when interfacing
with Apple specific API's since some of the API's are only available
in a form requiring using Blocks.

Interfacing with Blocks is done by directly using the underlying ABI.
A Block basically consist of a struct containing a function pointer to
the body of the Block, any captured values and some metadata.
The ABI that is used is the Apple ABI [2].

[1] Language specification for Blocks:
http://clang.llvm.org/docs/BlockLanguageSpec.html

[2] http://clang.llvm.org/docs/Block-ABI-Apple.html